Moves language visitors to sub-packages.
